refactor: rename aicore package to tiny-aicore, fix token refresh and lock contention
- Rename packages/aicore/ → packages/tiny-aicore/ (import: tiny_aicore)
- Fix HTTP-under-lock: do OAuth2 POST outside the token lock, re-acquire only to write
- Fix token refresh: OpenAI/Anthropic/Embeddings adapters now use a custom httpx.Auth
  subclass (_AICoreAuth) that calls client.get_token() on every request instead of
  baking the token at construction time (avoids 401 after ~1hr expiry)
- Remove dead self._name in GeminiAdapter
- Remove redundant f-string in OpenAIEmbeddingsAdapter
- Fix requires-python: ==3.14.* → >=3.14
