+++ packages/tiny-aicore/src/tiny_aicore/__init__.py
@@ -1,0 +1,15 @@
+"""SAP AI Core client and LangChain model adapters."""
+
+from tiny_aicore.client import AICoreClient
+from tiny_aicore.models.anthropic import AnthropicAdapter
+from tiny_aicore.models.embeddings import OpenAIEmbeddingsAdapter
+from tiny_aicore.models.gemini import GeminiAdapter
+from tiny_aicore.models.openai import OpenAIAdapter
+
+__all__ = [
+    "AICoreClient",
+    "AnthropicAdapter",
+    "GeminiAdapter",
+    "OpenAIAdapter",
+    "OpenAIEmbeddingsAdapter",
+]
