+++ packages/tiny-aicore/src/tiny_aicore/client.py
@@ -22,32 +22,37 @@ class AICoreClient:
         self._client_secret: str = config("AICORE_CLIENT_SECRET")
         self._resource_group: str = config("AICORE_RESOURCE_GROUP", default="default")
 
-        self._lock = threading.Lock()
+        self._token_lock = threading.Lock()
         self._token: str = ""
         self._token_exp: float = 0.0
+        self._deployment_lock = threading.Lock()
         self._deployment_cache: dict[str, str] = {}
 
     def get_token(self) -> str:
         """Return a cached or freshly fetched OAuth2 bearer token."""
-        with self._lock:
+        with self._token_lock:
             if self._token and time.time() < self._token_exp:
                 return self._token
 
-            resp = httpx.post(
-                f"{self._auth_url}/oauth/token",
-                data={"grant_type": "client_credentials"},
-                auth=(self._client_id, self._client_secret),
-                headers={"Content-Type": "application/x-www-form-urlencoded"},
-            )
-            resp.raise_for_status()
-            body = resp.json()
-            self._token = body["access_token"]
-            self._token_exp = time.time() + body["expires_in"] - 60
+        resp = httpx.post(
+            f"{self._auth_url}/oauth/token",
+            data={"grant_type": "client_credentials"},
+            auth=(self._client_id, self._client_secret),
+            headers={"Content-Type": "application/x-www-form-urlencoded"},
+        )
+        resp.raise_for_status()
+        body = resp.json()
+        token = body["access_token"]
+        token_exp = time.time() + body["expires_in"] - 60
+
+        with self._token_lock:
+            self._token = token
+            self._token_exp = token_exp
             return self._token
 
     def get_deployment_url(self, deployment_id: str) -> str:
         """Return the deploymentUrl for the given deployment ID (cached)."""
-        with self._lock:
+        with self._deployment_lock:
             if deployment_id in self._deployment_cache:
                 return self._deployment_cache[deployment_id]
 
@@ -65,6 +70,6 @@ class AICoreClient:
         if not url:
             raise ValueError(f"Deployment {deployment_id!r} has no deploymentUrl (not running?)")
 
-        with self._lock:
+        with self._deployment_lock:
             self._deployment_cache[deployment_id] = url
         return url
