+++ packages/tiny-aicore/src/tiny_aicore/models/openai.py
@@ -1,0 +1,41 @@
+"""OpenAI chat model adapter for SAP AI Core."""
+
+from collections.abc import Generator
+from typing import Any
+
+import httpx
+from langchain_openai import ChatOpenAI
+
+from tiny_aicore.client import AICoreClient
+
+
+class _AICoreAuth(httpx.Auth):
+    """httpx Auth that injects a fresh AI Core bearer token on every request."""
+
+    def __init__(self, client: AICoreClient) -> None:
+        self._client = client
+
+    def auth_flow(self, request: httpx.Request) -> Generator[httpx.Request, httpx.Response, None]:
+        request.headers["Authorization"] = f"Bearer {self._client.get_token()}"
+        yield request
+
+
+class OpenAIAdapter:
+    """Wraps ChatOpenAI with SAP AI Core deployment URL and per-request token refresh."""
+
+    def __init__(self, client: AICoreClient, deployment_id: str, model_name: str, **kwargs: Any) -> None:
+        deployment_url = client.get_deployment_url(deployment_id)
+        auth = _AICoreAuth(client)
+        self._llm = ChatOpenAI(
+            base_url=f"{deployment_url}/chat/completions",
+            api_key="placeholder",  # type: ignore[arg-type]  # auth header injected by transport
+            model=model_name,
+            http_client=httpx.Client(auth=auth),
+            http_async_client=httpx.AsyncClient(auth=auth),
+            **kwargs,
+        )
+
+    @property
+    def llm(self) -> ChatOpenAI:
+        """Return the underlying ChatOpenAI instance."""
+        return self._llm
